Move module registration and macro expansion to semantics
