--- src/__tests__/fixtures/index.ts
@@ -1,4 +1,0 @@
-export * from "../../parser/__tests__/fixtures/raw-parser-ast.js";
-export * from "../../parser/__tests__/fixtures/parse-text-void-file.js";
-export { regularMacrosAst } from "./regular-macros-ast.js";
-export { regularMacrosVoidFile } from "./regular-macros-void-file.js";
