+++ src/semantics/__tests__/fixtures/regular-macros-ast.ts
