+++ src/semantics/__tests__/fixtures/regular-macros-void-file.ts
@@ -1,5 +1,3 @@
-import { CharStream } from "../../parser/char-stream.js";
-
 export const regularMacrosVoidFile = `
 macro \`()
   quote quote $@body
