+++ src/semantics/__tests__/regular-macros.test.ts
@@ -1,10 +1,11 @@
 import { describe, it } from "node:test";
-import { parse } from "../parser/parser.js";
-import { regularMacrosAst, regularMacrosVoidFile } from "./fixtures/index.js";
+import { parse } from "../../parser/parser.js";
 import assert from "node:assert";
 import path from "node:path";
 import { registerModules } from "../modules.js";
 import { expandRegularMacros } from "../regular-macros.js";
+import { regularMacrosVoidFile } from "./fixtures/regular-macros-void-file.js";
+import { regularMacrosAst } from "./fixtures/regular-macros-ast.js";
 
 describe("regular macro evaluation", () => {
   it("should evaluate macros of the example file", async () => {
