+++ src/semantics/index.ts
@@ -1,13 +1,20 @@
 import { checkTypes } from "./check-types.js";
 import { initPrimitiveTypes } from "./init-primitive-types.js";
 import { initEntities } from "./init-entities.js";
-import { TypeChecker } from "./types.js";
+import { SemanticProcessor } from "./types.js";
+import { registerModules } from "./modules.js";
+import { expandRegularMacros } from "./regular-macros.js";
+import { ParsedModule } from "../parser/index.js";
+import { Expr } from "../syntax-objects/expr.js";
 
-const typePhases: TypeChecker[] = [
+const semanticPhases: SemanticProcessor[] = [
+  expandRegularMacros,
   initPrimitiveTypes,
   initEntities,
   checkTypes,
 ];
 
-export const typeCheck: TypeChecker = (expr) =>
-  typePhases.reduce((expr, checker) => checker(expr), expr);
+export const processSemantics = (parsedModule: ParsedModule): Expr => {
+  const expr = registerModules(parsedModule);
+  return semanticPhases.reduce((e, checker) => checker(e), expr as Expr);
+};
