+++ src/semantics/init-primitive-types.ts
@@ -1,7 +1,7 @@
 import { i32, f32, i64, f64, bool, dVoid } from "../syntax-objects/types.js";
-import { TypeChecker } from "./types.js";
+import { SemanticProcessor } from "./types.js";
 
-export const initPrimitiveTypes: TypeChecker = (expr) => {
+export const initPrimitiveTypes: SemanticProcessor = (expr) => {
   expr.registerEntity(i32);
   expr.registerEntity(f32);
   expr.registerEntity(i64);
