+++ src/semantics/types.ts
@@ -1,3 +1,3 @@
 import { Expr } from "../syntax-objects/expr.js";
 
-export type TypeChecker = (expr: Expr) => Expr;
+export type SemanticProcessor = (expr: Expr) => Expr;
